[search] Add builder and parser function.
