+++ search/search-action/index.js
@@ -2,83 +2,17 @@
 let assign = require('object-assign');
 let keys = require('lodash/object/keys');
 let {clone, isEqual} = require('lodash/lang');
-
+let _builder = require('./builder');
 const ALL = 'ALL';
 
-let _buildPagination = (pageInfos) => {
-    return {
-        page: pageInfos.page || 0,
-        skip: pageInfos.skip || 0
-    }
-};
-let _buildOrderAndSort = (pageInfos) => {
-    if (pageInfos.order) {
-        let result = {};
-        result.sortFieldName = pageInfos.order.key;
-        if (pageInfos.order.order) {
-            result.sortDesc = pageInfos.order.order.toLowerCase() === 'desc';
-        }
-        return result;
-    } else {
-        return {
-            sortFieldName: '',
-            sortDesc: false
-        }
-    }
-};
-let _buildFacets = (facets) => {
-    return keys(facets).map((selectedFacetKey) => {
-        let selectedFacet = facets[selectedFacetKey];
-        return {
-            key: selectedFacetKey,
-            value: selectedFacet.key
-        };
-    });
-};
-let _parseUnscopedResponse = (data, context) => {
-    return ({
-        map: data.groups,
-        facet: _parseFacets(data.facets),
-        pageInfos: _parsePageInfos(data, context)
-    });
-};
 
-let _parseScopedResponse = (data, context) => {
-    return ({
-        map: data.groups || {[context.scope]: data.list},
-        facet: _parseFacets(data.facets),
-        pageInfos: _parsePageInfos(data, context)
-    });
-};
-let _parseFacets = (facets) => {
-    return keys(facets).reduce((formattedFacets, serverFacetKey) => {
-        let serverFacetData = facets[serverFacetKey];
-        formattedFacets[serverFacetKey] = keys(serverFacetData).reduce((facetData, serverFacetItemKey) => {
-            let serverFacetItemValue = serverFacetData[serverFacetItemKey];
-            facetData[serverFacetItemKey] = {
-                label: serverFacetItemKey,
-                count: serverFacetItemValue
-            };
-            return facetData;
-        }, {});
-        return formattedFacets;
-    }, {});
-};
+
 
 let _dispatchResult = (data) => {
     Focus.dispatcher.handleServerAction({
         data,
         type: 'update'
     });
-};
-
-
-let _parsePageInfos = (data, context) => {
-    return {
-        currentPage: context.page,
-        perPage: 50,
-        totalRecords: data.totalCount
-    }
 };
 
 module.exports = function(config){
@@ -110,12 +44,18 @@ module.exports = function(config){
    * @return {function} - The builded search action.
    */
   return function searchAction(options){
-    options = options || {};
+    options = clone(options || {});
+
 
     ///////// COMMENT //////////
     // Could be nice to read all the info from the stores, so that no complex option needs to be passed to mutliple functions....
 
     let {scope, query} = options.criteria; //Cannot be read from the store ?
+    if(query === ''){
+      query = '*';
+    }
+    let urlData = assign(_builder.pagination(options.pageInfos), _builder.orderAndSort(options.pageInfos));
+
     if(scope === ALL){
       //Call the search action.
       options.service.scope(options);
